refactor: VCに残っていたビジネスロジックの分離、前回修正ができていなかった部分の修正
